requirements.txt
streamlit==1.31.0
transformers==4.37.2
torch==2.1.2
pyyaml==6.0.1
google-generativeai>=0.3.0